requirements.txt
Django==2.0.10
psycopg2>=2.7,<3.0
mysqlclient==1.3.13
mysql-connector-python==8.0.16
Pillow==5.4.1
python-decouple==3.1
django-mysql==2.2.0
